Only check items for workshedability
If you don't have a workshed item right now and you don't have a specific familiar, the function would return True, since the current workshed name (an empty string) is equal to the familiar name (an empty string). This fix makes it so only items can be haveUnrestricted on the premise of matching names.