A band-aid for the Monster component
There is a small window where Mafia says the location contains monsters, but cannot yet provide appearance rates. So all appearance rates are undefined, and we happily create entries for boss monsters and the like. And then they stay orphaned.

This is probably a general problem with the keys, and will break if you do a banish. Someone who understands React better than me should probably look at the code of that component.